Added tests for compiled contracts
