+++ .gitignore
@@ -7,7 +7,6 @@
 **/*.o
 **/*.asm
 output
-compiled
 
 ethtests
 *.tar.gz
