+++ compiled/Fibonacci.bin
@@ -1,0 +1,1 @@
+608060405234801561000f575f5ffd5b50610020600a61002a60201b60201c565b5f81905550610146565b5f5f820361003a575f9050610091565b6001820361004b5760019050610091565b5f5f90505f600190505f5f600290505b85811161008957828461006e91906100cc565b91508293508192508080610081906100ff565b91505061005b565b508093505050505b919050565b5f819050919050565b7f4e487b71000000000000000000000000000000000000000000000000000000005f52601160045260245ffd5b5f6100d682610096565b91506100e183610096565b92508282019050808211156100f9576100f861009f565b5b92915050565b5f61010982610096565b91507fffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff820361013b5761013a61009f565b5b600182019050919050565b61025e806101535f395ff3fe608060405234801561000f575f5ffd5b5060043610610034575f3560e01c806361047ff4146100385780636537214714610068575b5f5ffd5b610052600480360381019061004d919061012e565b610086565b60405161005f9190610168565b60405180910390f35b6100706100f2565b60405161007d9190610168565b60405180910390f35b5f5f8203610096575f90506100ed565b600182036100a757600190506100ed565b5f5f90505f600190505f5f600290505b8581116100e55782846100ca91906101ae565b915082935081925080806100dd906101e1565b9150506100b7565b508093505050505b919050565b5f5481565b5f5ffd5b5f819050919050565b61010d816100fb565b8114610117575f5ffd5b50565b5f8135905061012881610104565b92915050565b5f60208284031215610143576101426100f7565b5b5f6101508482850161011a565b91505092915050565b610162816100fb565b82525050565b5f60208201905061017b5f830184610159565b92915050565b7f4e487b71000000000000000000000000000000000000000000000000000000005f52601160045260245ffd5b5f6101b8826100fb565b91506101c3836100fb565b92508282019050808211156101db576101da610181565b5b92915050565b5f6101eb826100fb565b91507fffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff820361021d5761021c610181565b5b60018201905091905056fea2646970667358221220cc7e80883e46c9b1dcc50e00d58d2ebc200d477cea82d9c2fc35d1ee0999b9ce64736f6c634300081b0033
