+++ tests/evm.rs
@@ -5641,25 +5641,3 @@ fn refund_limit_value() {
 
     run_program_assert_gas_and_refund(env, db, needed_gas as _, used_gas as _, refunded_gas as _);
 }
-
-#[test]
-fn factorial_binary() {
-    let address = Address::from_low_u64_be(3000);
-    // binary generated of factorial.sol, calculates 10! and put it in storage
-    let initialization_code = hex::decode("608060405234801561000f575f5ffd5b50610020600a61002a60201b60201c565b5f81905550610126565b5f5f600190505f600290505b83811161005e578082610049919061009e565b91508080610056906100df565b915050610036565b5080915050919050565b5f819050919050565b7f4e487b71000000000000000000000000000000000000000000000000000000005f52601160045260245ffd5b5f6100a882610068565b91506100b383610068565b92508282026100c181610068565b915082820484148315176100d8576100d7610071565b5b5092915050565b5f6100e982610068565b91507fffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff820361011b5761011a610071565b5b600182019050919050565b61023e806101335f395ff3fe608060405234801561000f575f5ffd5b5060043610610034575f3560e01c806365372147146100385780638371483414610056575b5f5ffd5b610040610086565b60405161004d91906100e1565b60405180910390f35b610070600480360381019061006b9190610128565b61008b565b60405161007d91906100e1565b60405180910390f35b5f5481565b5f5f600190505f600290505b8381116100bf5780826100aa9190610180565b915080806100b7906101c1565b915050610097565b5080915050919050565b5f819050919050565b6100db816100c9565b82525050565b5f6020820190506100f45f8301846100d2565b92915050565b5f5ffd5b610107816100c9565b8114610111575f5ffd5b50565b5f81359050610122816100fe565b92915050565b5f6020828403121561013d5761013c6100fa565b5b5f61014a84828501610114565b91505092915050565b7f4e487b71000000000000000000000000000000000000000000000000000000005f52601160045260245ffd5b5f61018a826100c9565b9150610195836100c9565b92508282026101a3816100c9565b915082820484148315176101ba576101b9610153565b5b5092915050565b5f6101cb826100c9565b91507fffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff82036101fd576101fc610153565b5b60018201905091905056fea264697066735822122003c705469d87a20b4404d75e8e3f61894bbc76a3b629a315e16820b749a5f27864736f6c634300081b0033").unwrap();
-
-    let bytes = Bytes::from(initialization_code);
-
-    let db = Db::new().with_contract(address, bytes);
-    let mut env = Env::default();
-    env.tx.gas_limit = 999_999;
-    env.tx.transact_to = TransactTo::Call(address);
-    let mut evm = Evm::new(env, db);
-    let result = evm.transact().unwrap();
-    assert!(result.result.is_success());
-    let state = result.state.get(&address).unwrap();
-    assert_eq!(
-        state.storage.get(&EU256::zero()).unwrap().present_value,
-        ethereum_types::U256::from(3628800)
-    )
-}
